Estimate treatment coverage

Estimate treatment coverage

Usage

calculate_tc(cin, cout, rin, k = 3)

Arguments

cin

Cases in CMAM programme

cout

Cases not in CMAM programme

rin

Recovering cases in CMAM programme

k

Correction factor. Ratio of the mean length of an untreated episode to the mean length of a CMAM treatment episode

Value

Value of treatment coverage

References

Safari Balegamire, Katja Siling, Jose Luis Alvarez Moran, Ernest Guevarra, Sophie Woodhead, Alison Norris, Lionella Fieschi, Paul Binns, and Mark Myatt (2015). A single coverage estimator for use in SQUEAC, SLEAC, and other CMAM coverage assessments. Field Exchange 49, March 2015. p81. <www.ennonline.net/fex/49/singlecoverage>

Author

Ernest Guevarra based on technical notes and equations by Mark Myatt

Examples

calculate_tc(cin = 5, cout = 20, rin = 5, k = 3)
#> [1] 0.2857143